refactor(backend): move export function
